+++ src/issues/SemanticValidationIssues.ts
@@ -2,9 +2,21 @@ import { ValidationIssue } from "../validation/ValidationIssue";
 import { ValidationIssueSeverity } from "../validation/ValidationIssueSeverity";
 import { ValidationIssueUtils } from "./ValidationIssueUtils";
 
-// TODO Each of these issues should be documented,
-// even more extensively than the basic ones!
+/**
+ * Methods to create `ValidationIssue` instances that describe
+ * issues related to the semantical validity of tilesets.
+ */
 export class SemanticValidationIssues {
+  /**
+   * Indicates that the 'version' string of a tileset asset
+   * had an unknown value.
+   *
+   * (The known values are defined by the 'AssetValidator' class)
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static ASSET_VERSION_UNKNOWN(path: string, message: string) {
     const type = "ASSET_VERSION_UNKNOWN";
     const severity = ValidationIssueSeverity.WARNING;
@@ -12,6 +24,18 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates that the 'refine' value of a tile was valid, but
+   * had an unexpected case.
+   *
+   * This only a warning, intended for legacy tilesets, where a
+   * value like 'Replace' was still valid. Current tilesets should
+   * always use uppercase values like 'REPLACE'.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static TILE_REFINE_WRONG_CASE(path: string, message: string) {
     const type = "TILE_REFINE_WRONG_CASE";
     const severity = ValidationIssueSeverity.WARNING;
@@ -19,6 +43,23 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates that the root tile of an implicit tileset was invalid.
+   *
+   * This is caused by the root tile of an implicit tileset defining
+   * one of the properties that are disallowed for implicit roots:
+   * - tile.children
+   * - tile.metadata
+   * - tile.content.boundingVolume
+   *
+   * It may also indicate that the required subtree information
+   * could not be created (for example, when the subtree data
+   * could not be read)
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static TILE_IMPLICIT_ROOT_INVALID(path: string, message: string) {
     const type = "TILE_IMPLICIT_ROOT_INVALID";
     const severity = ValidationIssueSeverity.ERROR;
@@ -26,13 +67,51 @@ export class SemanticValidationIssues {
     return issue;
   }
 
-  static BOUNDING_VOLUME_INCONSISTENT(path: string, message: string) {
-    const type = "BOUNDING_VOLUME_INCONSISTENT";
+  /**
+   * Indicates that a single bounding volume was invalid.
+   *
+   * This refers to certain constraints that are applied to
+   * specific bounding volume types. For example, that the
+   * radius of a bounding sphere may not be negative, or
+   * that the borders of a bounding regions are within
+   * valid ranges.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
+  static BOUNDING_VOLUME_INVALID(path: string, message: string) {
+    const type = "BOUNDING_VOLUME_INVALID";
     const severity = ValidationIssueSeverity.ERROR;
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
 
+  /**
+   * Indicates that a bounding volume structure was inconsistent.
+   *
+   * For now, this only means that a content bounding volume was
+   * not fully contained in the tile bounding volume.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
+  static BOUNDING_VOLUMES_INCONSISTENT(path: string, message: string) {
+    const type = "BOUNDING_VOLUMES_INCONSISTENT";
+    const severity = ValidationIssueSeverity.ERROR;
+    const issue = new ValidationIssue(type, path, message, severity);
+    return issue;
+  }
+
+  /**
+   * Indicates that the minimum value of a 'tileset.properties'
+   * element was larger than the maximum.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static PROPERTIES_MINIMUM_LARGER_THAN_MAXIMUM(path: string, message: string) {
     const type = "PROPERTIES_MINIMUM_LARGER_THAN_MAXIMUM";
     const severity = ValidationIssueSeverity.ERROR;
@@ -40,211 +119,33 @@ export class SemanticValidationIssues {
     return issue;
   }
 
-  static TILE_GEOMETRIC_ERROR_INCONSISTENT(path: string, message: string) {
-    const type = "TILE_GEOMETRIC_ERROR_INCONSISTENT";
+  /**
+   * Indicates that the geometric error of a tile was larger than
+   * the geometric error of its parent.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
+  static TILE_GEOMETRIC_ERRORS_INCONSISTENT(path: string, message: string) {
+    const type = "TILE_GEOMETRIC_ERRORS_INCONSISTENT";
     const severity = ValidationIssueSeverity.ERROR;
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
 
-  static CLASS_PROPERTY_COMPONENT_TYPE_WITH_INVALID_TYPE(
-    path: string,
-    componentType: string,
-    theType: string
-  ) {
-    const type = "CLASS_PROPERTY_COMPONENT_TYPE_WITH_INVALID_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The 'componentType' was defined to be '${componentType}', but ` +
-      `must be undefined for a property with type '${theType}'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_COMPONENT_TYPE_MISSING(path: string, theType: string) {
-    const type = "CLASS_PROPERTY_COMPONENT_TYPE_MISSING";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The 'componentType' must be defined for a ` +
-      `property with type '${theType}'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_ENUMTYPE_WITH_NON_ENUM_TYPE(
-    path: string,
-    enumType: string,
-    theType: string
-  ) {
-    const type = "CLASS_PROPERTY_ENUMTYPE_WITH_NON_ENUM_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The 'enumType' was defined to be '${enumType}', but ` +
-      `must be undefined for a property with type '${theType}'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_ENUM_TYPE_WITHOUT_ENUMTYPE(path: string) {
-    const type = "CLASS_PROPERTY_ENUM_TYPE_WITHOUT_ENUMTYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message = `The property has the type 'ENUM', but no 'enumType' was defined`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_ENUMTYPE_NOT_FOUND(
-    path: string,
-    propertyName: string,
-    enumType: string
-  ) {
-    const type = "CLASS_PROPERTY_ENUMTYPE_NOT_FOUND";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' refers to the 'enumType' ` +
-      `'${enumType}', but the schema does not define this 'enumType'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_VALUE_ENUM_VALUE_NOT_FOUND(
-    path: string,
-    name: string,
-    propertyName: string,
-    enumType: string | undefined,
-    enumValueName: string
-  ) {
-    const type = "CLASS_PROPERTY_ENUM_VALUE_NOT_FOUND";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The value '${name}' of property '${propertyName}' refers to a value ` +
-      `with the name '${enumValueName}' of the enum '${enumType}', but this ` +
-      `enum does not define a value with this name`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_COUNT_FOR_NON_ARRAY(
-    path: string,
-    propertyName: string
-  ) {
-    const type = "CLASS_PROPERTY_COUNT_FOR_NON_ARRAY";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' defines a 'count', but ` +
-      `the property is not an array`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_NORMALIZED_FOR_NON_NORMALIZABLE_TYPE(
-    path: string,
-    propertyName: string,
-    propertyType: string
-  ) {
-    const type = "CLASS_PROPERTY_NORMALIZED_FOR_NON_NORMALIZABLE_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' is defined to be 'normalized', ` +
-      `but the type '${propertyType}' can not be normalized`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_NORMALIZED_FOR_NON_INTEGER_COMPONENT_TYPE(
-    path: string,
-    propertyName: string,
-    componentType: string
-  ) {
-    const type = "CLASS_PROPERTY_NORMALIZED_FOR_NON_NORMALIZABLE_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' is defined to be 'normalized', ` +
-      `but the component type '${componentType}' is not an integer type`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_OFFSET_SCALE_FOR_NON_FLOATING_POINT_TYPE(
-    path: string,
-    propertyName: string,
-    offsetOrScale: string,
-    propertyType: string,
-    componentType: string | undefined,
-    normalized: boolean | undefined
-  ) {
-    const type = "CLASS_PROPERTY_OFFSET_SCALE_FOR_NON_FLOATING_POINT_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' is defines '${offsetOrScale}', ` +
-      `which is only applicable to properties with types 'SCALAR', ` +
-      `'VEC2', 'VEC3', 'VEC4', 'MAT2', 'MAT3', or 'MAT4' when they have ` +
-      `component types 'FLOAT32' or 'FLOAT64', or when they are normalized ` +
-      `and have component types 'INT8', 'UINT8', 'INT16', 'UINT16', 'INT32', ` +
-      `'UINT32', 'INT64', or 'UINT64', but the property has type ` +
-      `'${propertyType}' with component type '${componentType}' and ` +
-      `'normalized' is '${normalized}`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTY_MIN_MAX_FOR_NON_NUMERIC_TYPE(
-    path: string,
-    propertyName: string,
-    minOrMax: string,
-    propertyType: string
-  ) {
-    const type = "CLASS_PROPERTY_MIN_MAX_FOR_NON_NUMERIC_TYPE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The property '${propertyName}' is defines '${minOrMax}', ` +
-      `which is only applicable to properties with types 'SCALAR', ` +
-      `'VEC2', 'VEC3', 'VEC4', 'MAT2', 'MAT3', or 'MAT4', but the ` +
-      `property has type '${propertyType}'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  // TODO Some of the highly specific issues above could be
-  // summarized in this one, with helpful messages...
-  static CLASS_PROPERTY_TYPE_ERROR(path: string, message: string) {
-    const type = "CLASS_PROPERTY_TYPE_ERROR";
-    const severity = ValidationIssueSeverity.ERROR;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static CLASS_PROPERTIES_DUPLICATE_SEMANTIC(
-    path: string,
-    propertyNameA: string,
-    propertyNameB: string,
-    semantic: string
-  ) {
-    const type = "CLASS_PROPERTIES_DUPLICATE_SEMANTIC";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message =
-      `The semantic '${semantic}' was assigned to property ` +
-      `'${propertyNameA}' and property '${propertyNameB}'`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static ENUM_VALUE_DUPLICATE_NAME(path: string, name: string) {
-    const type = "ENUM_VALUE_DUPLICATE_NAME";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message = `There enum value name '${name}' is not unique`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
-  static ENUM_VALUE_DUPLICATE_VALUE(path: string, value: number) {
-    const type = "ENUM_VALUE_DUPLICATE_VALUE";
-    const severity = ValidationIssueSeverity.ERROR;
-    const message = `There enum value '${value}' is not unique`;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
+  /**
+   * Indicates that a template URI contained an invalid variable name.
+   *
+   * The template URIs that are used for contents or subtree files
+   * in implicit tiling may contain variables like '{level}', and
+   * this issue indicates that there was an invalid variable name.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param variableName The variable name
+   * @param validVariableNames The valid variable names
+   * @returns The `ValidationIssue`
+   */
   static TEMPLATE_URI_INVALID_VARIABLE_NAME(
     path: string,
     variableName: string,
@@ -259,6 +160,19 @@ export class SemanticValidationIssues {
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
+
+  /**
+   * Indicates that a template URI did not contain an expected variable name.
+   *
+   * The template URIs that are used for contents or subtree files
+   * in implicit tiling are expected to contain certain variable
+   * names. This issue is only a 'WARNING' for the case that an
+   * expected name was not used.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param missingVVariableNames The missing variable names
+   * @returns The `ValidationIssue`
+   */
   static TEMPLATE_URI_MISSING_VARIABLE_NAME(
     path: string,
     missingVVariableNames: string[]
@@ -274,13 +188,21 @@ export class SemanticValidationIssues {
     return issue;
   }
 
-  static TRAVERSAL_ERROR(path: string, message: string) {
-    const type = "TRAVERSAL_ERROR";
-    const severity = ValidationIssueSeverity.ERROR;
-    const issue = new ValidationIssue(type, path, message, severity);
-    return issue;
-  }
-
+  /**
+   * Indicates an error in an implicit tileset structure.
+   *
+   * This is a generic error indicating that the internal structures
+   * for traversing the implicit tileset could not be created.
+   * Clients should rarely see this message, because errors that
+   * prevent the traversal should be caught earlier (and prevent
+   * the traversal attempts). But if it happens, the 'message'
+   * should contain further information about the reason for
+   * the error.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static IMPLICIT_TILING_ERROR(path: string, message: string) {
     const type = "IMPLICIT_TILING_ERROR";
     const severity = ValidationIssueSeverity.ERROR;
@@ -288,6 +210,17 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates an inconsistency of buffers and buffer views.
+   *
+   * This mainly refers to the 'buffers' and 'bufferViews' of
+   * an implicit subtree. It may, for example, inciate that a
+   * buffer view does not fit into the buffer that it refers to.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static BUFFERS_INCONSISTENT(path: string, message: string) {
     const type = "BUFFERS_INCONSISTENT";
     const severity = ValidationIssueSeverity.ERROR;
@@ -295,6 +228,22 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates an inconsistency in availability information.
+   *
+   * The availability information for tiles, content, and child
+   * subtrees that is stored as part of an implicit tileset has
+   * to obey certain constraints. For example:
+   * - When a content is available, then the tile must be available
+   * - When a tile is available, then the parent tile must be available
+   *
+   * More specific information about the inconsistency is given
+   * in the error message.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static SUBTREE_AVAILABILITY_INCONSISTENT(path: string, message: string) {
     const type = "SUBTREE_AVAILABILITY_INCONSISTENT";
     const severity = ValidationIssueSeverity.ERROR;
@@ -302,6 +251,17 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates that a tile transform was invalid.
+   *
+   * The exact constraints for being 'valid' are not specified.
+   * For now, this indicates that the transform matrix was
+   * the zero-matrix.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param message The message for the `ValidationIssue`
+   * @returns The `ValidationIssue`
+   */
   static TRANSFORM_INVALID(path: string, message: string) {
     const type = "TRANSFORM_INVALID";
     const severity = ValidationIssueSeverity.ERROR;
@@ -309,6 +269,15 @@ export class SemanticValidationIssues {
     return issue;
   }
 
+  /**
+   * Indicates that a certain extension was listed in the
+   * 'extensionsRequired', but not in the 'extensionsUsed'
+   * of a tileset.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param extensionName The extension name
+   * @returns The `ValidationIssue`
+   */
   static EXTENSION_REQUIRED_BUT_NOT_USED(path: string, extensionName: string) {
     const type = "EXTENSION_REQUIRED_BUT_NOT_USED";
     const severity = ValidationIssueSeverity.ERROR;
@@ -318,6 +287,19 @@ export class SemanticValidationIssues {
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
+
+  /**
+   * Indicates that a certain extension was found, but was not
+   * listed in the 'extensionsUsed' of a tileset.
+   *
+   * An extension is 'found' when it is encountered in any
+   * 'someRootProperty.extensons' dictionary during the
+   * traversal,
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param extensionName The extension name
+   * @returns The `ValidationIssue`
+   */
   static EXTENSION_FOUND_BUT_NOT_USED(path: string, extensionName: string) {
     const type = "EXTENSION_FOUND_BUT_NOT_USED";
     const severity = ValidationIssueSeverity.ERROR;
@@ -327,6 +309,23 @@ export class SemanticValidationIssues {
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
+
+  /**
+   * Indicates that a certain extension was listed in the
+   * 'extensionsUsed' of a tileset, but not found during
+   * the traversal.
+   *
+   * An extension is 'found' when it is encountered in any
+   * 'someRootProperty.extensons' dictionary during the
+   * traversal.
+   *
+   * NOTE: The exact mechanism for an extension being "used" may have to be
+   * reviewed. See https://github.com/CesiumGS/3d-tiles-validator/issues/231
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param extensionName The extension name
+   * @returns The `ValidationIssue`
+   */
   static EXTENSION_USED_BUT_NOT_FOUND(path: string, extensionName: string) {
     const type = "EXTENSION_USED_BUT_NOT_FOUND";
     const severity = ValidationIssueSeverity.WARNING;
@@ -336,6 +335,16 @@ export class SemanticValidationIssues {
     const issue = new ValidationIssue(type, path, message, severity);
     return issue;
   }
+
+  /**
+   * Indicates that a certain extension was found during
+   * the traversal, but is not known or handled by the
+   * validator in any way.
+   *
+   * @param path The JSON path for the `ValidationIssue`
+   * @param extensionName The extension name
+   * @returns The `ValidationIssue`
+   */
   static EXTENSION_NOT_SUPPORTED(path: string, extensionName: string) {
     const type = "EXTENSION_NOT_SUPPORTED";
     const severity = ValidationIssueSeverity.WARNING;
