+++ src/validation/metadata/BinaryPropertyTableValidator.ts
@@ -621,7 +621,7 @@ export class BinaryPropertyTableValidator {
         `have a byteLength of ${numValues}*${componentTypeByteSize} = ` +
         `${expectedByteLength}. But the buffer view with index ` +
         `${bufferViewIndex} has a byteLength of ${actualByteLength}`;
-      const issue = MetadataValidationIssues.METADATA_INVALID_SIZE(
+      const issue = MetadataValidationIssues.METADATA_INVALID_LENGTH(
         path,
         message
       );
@@ -677,7 +677,7 @@ export class BinaryPropertyTableValidator {
         numValues,
         expectedByteLength
       );
-      const issue = MetadataValidationIssues.METADATA_INVALID_SIZE(
+      const issue = MetadataValidationIssues.METADATA_INVALID_LENGTH(
         path,
         message
       );
